Compare podif status and move podif code to new file
